moved profile search bar to homepage fragment
