+++ code/app/src/main/java/com/github/bytebandits/bithub/view/ProfileFragment.java
@@ -5,22 +5,16 @@ import android.os.Handler;
 import android.os.Looper;
 import android.util.Log;
 import android.view.LayoutInflater;
-import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ImageButton;
-import android.widget.ImageView;
 import android.widget.ListView;
-import android.widget.SearchView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.fragment.app.Fragment;
 
-import com.github.bytebandits.bithub.MainActivity;
 import com.github.bytebandits.bithub.controller.DatabaseManager;
 import com.github.bytebandits.bithub.model.MoodPost;
 import com.github.bytebandits.bithub.R;
@@ -29,13 +23,7 @@ import com.github.bytebandits.bithub.model.Profile;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 
-import org.json.JSONException;
-import org.json.JSONObject;
-
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -185,8 +173,6 @@ public class ProfileFragment extends Fragment {
             }
         });
 
-        profileSearchManager(view);
-
         return view;
     }
 
@@ -197,88 +183,4 @@ public class ProfileFragment extends Fragment {
         SettingsDialog settingsDialog = new SettingsDialog(requireContext());
         settingsDialog.showSettingsDialog();
     }
-
-    /**
-     * Manages and encapsulates all logic relating to searching profiles within the profile fragment
-     * @param view the view in question, so the method can reference the various UI elements of the layout
-     */
-private void profileSearchManager(View view) {
-    ArrayList<Profile> profiles = new ArrayList<>();
-    ProfileSearchAdapter profileSearchAdapter = new ProfileSearchAdapter(requireContext(), profiles);
-
-    SearchView profileSearch = view.findViewById(R.id.profileSearch);
-    ListView profileResults = view.findViewById(R.id.profileResults);
-    profileResults.setAdapter(profileSearchAdapter);
-
-    // logic of profile transition via clicking
-    profileResults.setOnItemClickListener((adapterView, view1, i, l) -> {
-        Profile profile = profiles.get(i);
-        Toast.makeText(requireContext(), "Clicked: " + profile.getUserID(), Toast.LENGTH_SHORT).show();
-        ProfileFragment profileFragment = new ProfileFragment();
-        profileFragment.setIsOtherProfile(true);
-        profileFragment.setOtherProfile(profile);
-        ((MainActivity) requireActivity()).replaceFragment(profileFragment);
-    });
-
-    // logic to unfocus from search view, although there are cases where if you click certain ui elements, the click wont register and you will be still focused
-    ConstraintLayout parentConstraintLayout = view.findViewById(R.id.constraintLayoutProfile);
-    parentConstraintLayout.setOnClickListener(new View.OnClickListener() {
-        @Override
-        public void onClick(View view) {
-            profileSearch.clearFocus();
-        }
-    });
-
-    // submit logic for search results
-    ImageButton profileSearchIcon = view.findViewById(R.id.profileSearchConfirm);
-    profileSearchIcon.setOnClickListener(new View.OnClickListener(){
-        @Override
-        public void onClick(View view) {
-            String query = profileSearch.getQuery().toString();
-            if (!query.isEmpty()){
-                profileResults.setVisibility(View.VISIBLE);
-                DatabaseManager.getInstance().searchUsers(query, users -> {
-                    profiles.clear();
-
-                    for (HashMap<String, Object> user : users) {
-                        String profileJson = (String) user.get("profile");
-                        try{
-                            JSONObject profileObj = new JSONObject(profileJson);
-                            String userID = profileObj.getString("userID");
-                            String userImg;
-                            if (profileObj.isNull("img")) {
-                                profiles.add(new Profile(userID));
-                                // use default profile pic
-                            }
-//                            else{
-//                                  // do stuff if profile pic exists
-//                            }
-                        }
-                        catch (JSONException e) {
-                            Log.e("ProfileSearchJSONException", e.toString());
-                        }
-                    }
-                    profileSearchAdapter.notifyDataSetChanged();
-                });
-            }
-        }
-    });
-
-    profileSearch.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-        @Override
-        public boolean onQueryTextSubmit(String s) {
-            return false;
-        }
-
-        @Override
-        public boolean onQueryTextChange(String s) {
-            if (s.isEmpty()) {
-                profiles.clear();
-                profileSearchAdapter.notifyDataSetChanged();
-                profileResults.setVisibility(View.INVISIBLE);
-            }
-            return true;
-        }
-    });
-}
 }
