+++ code/app/src/main/res/layout/fragment_profile.xml
@@ -15,16 +15,6 @@
         android:focusable="auto"
         android:focusableInTouchMode="true">
 
-        <include
-            layout="@layout/profile_search_bar"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="8dp"
-            app:layout_constraintEnd_toStartOf="@id/filter_button"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
         <TextView
             android:id="@+id/textView2"
             android:layout_width="wrap_content"
